requirements.txt
attrs==17.2.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
gcloud==0.17.0
github3.py==0.9.6
googleapis-common-protos==1.5.2
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jws==0.1.3
MarkupSafe==0.23
oauth2client==3.0.0
oauthlib==2.0.2
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
python-twitter==3.3
requests==2.11.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.10.0
slackbot==0.4.1
slacker==0.9.50
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.21.1
websocket-client==0.42.1
Werkzeug==0.12.2
